requirements.txt
django == 1.9.5
django-guardian == 1.4.4
django-filter == 0.13.0
djangorestframework == 3.3.3
django-rest-swagger == 0.3.6
django-extensions == 1.6.1
netaddr == 0.7.18
django-cors-headers == 1.1.0
djangorestframework-filters == 0.8.0
mysql-python == 1.2.5
celery == 3.1.23
django-celery == 3.1.17
gunicorn == 19.4.5
redis == 2.10.5
django-cors-headers == 1.1.0